refactor(spatial): make the point-cloud ingest helpers private
`point_cloud.py` was a public module holding twenty private names and nine
public-looking ones, only one of which -- `ColumnDataType` -- was reachable
through `nominal.core`. The other eight could only be imported by reaching for
the module path, and six of them are the importer's own wire vocabulary, which no
caller has any use for.

- `point_cloud.py` -> `_point_cloud.py`, matching the six `_`-prefixed modules
  `nominal/core` already has.
- `GeometryType`, `Sampler`, `RealMeasurement`, `ScalarAttributeType`,
  `RealAttributeType` and `AttributeType` are private.
- `DEFAULT_RGB_ATTRIBUTE` and `DEFAULT_POINT_CLOUD_CHANNEL` are gone. They had
  one real use each, as defaults on a public signature, where they rendered as
  indirection in the docs; the signatures carry `"color"` and `"point_cloud"`
  directly. The explanation of why colour is one hex column moved next to the
  code that builds the Rgb attribute.
- `ColumnDataType` stays public and stays here: the classifier produces it, and
  moving it to `spatial.py` would make this module import that one and cycle.
  `nominal/core/__init__.py` imports it from here, as it already does for five
  other private modules.

The public surface for ingesting a point cloud is now four names, all reached
through `nominal.core`: `Spatial`, `PointCloudMetadata`, `ScanPattern` and
`ColumnDataType`.
